fix(auth): 修复用户 ID 超 int 范围导致的鉴权与查询失效
用户表主键由雪花算法生成（如 2091818653023719424），早已超出 int 范围。
JwtUtil/AuthFilter 改为 Long 后，下游仍有多处用 Integer 承载 userId：

- ChatUserController.get_currentUserId 按 `attr instanceof Integer` 取值，
  AuthFilter 写入 Long 后恒为 false，导致所有已登录用户被判成"未登录"
  （/user/info、头像上传、资料更新全部失效）——本次 P0 回归
- getUserById 参数改为 Long，去掉 intValue() 截断雪花 ID 的隐患
- /infoById 参数改为 Long，避免大 ID 用户在参数绑定阶段直接 400
- 离线消息投递 parseUserId 改用 Long.parseLong，修复大 ID 用户上线后
  收不到离线消息

补充 JwtUtilTest（覆盖超 int 范围 ID 解析）与 AuthFilterTest（覆盖
鉴权放行、缺 token、坏 token、白名单四类路径）锁定该回归。

+++ snow-chat-im-backend/snow-im-api/src/main/java/com/stylesmile/chat/mapper/ChatOfflineMessageMapper.java
@@ -16,10 +16,10 @@ public interface ChatOfflineMessageMapper extends BaseMapper<ChatOfflineMessage>
     /**
      * 查询接收方的离线消息
      */
-    List<ChatOfflineMessage> findByToUserId(@Param("toUserId") Integer toUserId);
+    List<ChatOfflineMessage> findByToUserId(@Param("toUserId") Long toUserId);
 
     /**
      * 删除接收方的离线消息
      */
-    void deleteByToUserId(@Param("toUserId") Integer toUserId);
+    void deleteByToUserId(@Param("toUserId") Long toUserId);
 }
